Uploaded FITS files and updates
Uploaded all of the FITS files that will be used for the final project in the form of zip files as they are too large to upload otherwise. I have changed a couple of things on the SEP tutorial and I have written the prototype for the UDF_105w object detection, but it still needs some work.